+++ arm/src/inst/smlald.c
@@ -16,8 +16,8 @@ int32 arm_inst_smlald(cpu_state_t *st, uint32 inst)
 	uint32 operand = (x == 1) ? ror(regv(rs), 16) : regv(rs);
 	int64 accvalue = (((uint64)regv(rdh)) << 32) | regv(rdl);
 
-	int32 product1 = (int16)bitm(regv(rm), 0, 15) * (int16)bitm(operand, 0, 15);
-	int32 product2 = (int16)bitm(regv(rm), 16, 31) * (int16)bitm(operand, 16, 31);
+	int32 product1 = (int32)(int16)bitm(regv(rm), 0, 15) * (int16)bitm(operand, 0, 15);
+	int32 product2 = (int32)(int16)bitm(regv(rm), 16, 31) * (int16)bitm(operand, 16, 31);
 
 	int64 result = accvalue + product1 + product2;
