+++ arm/src/inst/smmla.c
@@ -13,11 +13,11 @@ int32 arm_inst_smmla(cpu_state_t *st, uint32 inst)
 	uint32 rm = inst_b4(0);
 	uint32 r = inst_b1(5);
 
-	int64 value = (int32)regv(rm) * (int32)regv(rs);
+	int64 value = (int64)(int32)regv(rm) * (int32)regv(rs);
 	if (r == 1)
-		regv(rd) = (((uint64)regv(rn) << 32) + value + 0x80000000) >> 32;
+		regv(rd) = (((int64)(int32)regv(rn) << 32) + value + 0x80000000) >> 32;
 	else
-		regv(rd) = (((int64)regv(rn) << 32) + value) >> 32;
+		regv(rd) = (((int64)(int32)regv(rn) << 32) + value) >> 32;
 
 	return EXEC_SUCCESS;
 }
